fix warn

+++ utils/src/pip_util.cpp
@@ -95,14 +95,14 @@ void PiPUtil::GetRectByScale(const uint32_t width, const uint32_t height, const 
     }
 }
 
-bool PiPUtil::GetValidRect(const uint32_t width, const int32_t height, Rect& rect)
+bool PiPUtil::GetValidRect(const int32_t width, const int32_t height, Rect& rect)
 {
     int32_t safePaddingHorizontal = static_cast<int32_t>(SAFE_PADDING_HORIZONTAL_VP * g_vpr);
     bool hasChanged = false;
     if (rect.posX_ < safePaddingHorizontal) {
         rect.posX_ = safePaddingHorizontal;
         hasChanged = true;
-    } else if ((rect.posX_ + rect.width_) > (width - safePaddingHorizontal)) {
+    } else if ((rect.posX_ + static_cast<int32_t>(rect.width_)) > (width - safePaddingHorizontal)) {
         rect.posX_ = width - safePaddingHorizontal - rect.width_;
         hasChanged = true;
     }
